api/requirements.txt
django>=4.1.7,<4.2
djangorestframework>=3.14.0,<3.15
django-cors-headers>=3.14.0,<3.15
django-storages[google]>=1.13.2,<1.14
drf-spectacular>=0.25.1,<0.26
google-auth>=2.17.3,<2.18
google-cloud-run>=0.7.1,<0.8
google-cloud-tasks>=2.13.1,<2.14
grpcio>=1.54.0,<1.55
openai>=0.27.0,<0.28
psycopg2>=2.9.5,<2.10
uwsgi>=2.0.21,<2.1
retry>=0.9.2,<0.10